requirements.txt
streamlit
sqlalchemy
pydantic
google-generativeai
python-dotenv
bcrypt
pandas
numpy
plotly


